Basic collider with collision detection added to player
