+++ Assets/Scripts/SolidPlayer.cs
@@ -1,0 +1,38 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SolidPlayer : MonoBehaviour
+{
+    private SpawnManager spawnManager;
+    private PlayerController playerController;
+    // Start is called before the first frame update
+    void Start()
+    {
+        spawnManager = FindObjectOfType<SpawnManager>();
+        playerController = FindObjectOfType<PlayerController>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    // What the heck are we hitting?
+    void OnCollisionEnter(Collision collision)
+    {
+        if (collision.gameObject.CompareTag("Enemy"))
+        {
+            Debug.Log("Enemy hit");
+            Destroy(collision.gameObject);
+            spawnManager.numberOfEnemies--;
+        }
+        else if(collision.gameObject.CompareTag("Rest Point"))
+        {
+            Debug.Log("Refreshing energy");
+            Destroy(collision.gameObject);
+            playerController.ResetEnergy();
+        }
+    }
+}
